feat: 백엔드 서버 연동 및 로그인 시 닉네임 표시 기능 구현
