--- src/app/components/DummyLoginModal/DummyLoginModal.css
@@ -1,377 +1,0 @@
-.login-container .text-xs-center a {
-  color: #5cb85c;
-  text-decoration: none;
-}
-
-.login-container {
-  width: 389px;
-  margin: 0 auto;
-}
-
-.login-container .btn {
-  background-color: #5cb85c;
-  border-color: #5cb85c;
-}
-
-.find {
-  margin-right: 30px;
-  display: flex;
-  align-items: center;
-}
-
-.find > a {
-  margin: 10px;
-  color: #8c8c8c;
-  font-family: Noto Sans KR;
-  font-size: 14px;
-  letter-spacing: -0.333px;
-}
-
-.middleLine {
-  position: relative;
-  bottom: 2px;
-  margin: 0;
-  display: flex;
-  align-items: center;
-  color: #8c8c8c;
-  height: 10px;
-}
-
-.login-container h1 {
-  color: #252628;
-  font-family: Noto Sans KR;
-  font-size: 24px;
-  font-style: normal;
-  font-weight: 500;
-  line-height: normal;
-  letter-spacing: -0.333px;
-  margin-bottom: 40px;
-}
-
-.modal {
-  position: fixed;
-  top: 0;
-  right: 0;
-  bottom: 0;
-  left: 0;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  background-color: rgba(0, 0, 0, 0.6);
-}
-
-.modal .loginModal {
-  width: 522px;
-  height: 817px;
-  background-color: white;
-  position: relative;
-  box-sizing: border-box;
-  margin: 50px auto;
-  padding: 20px;
-  background: #fff;
-  border-radius: 6px;
-  box-shadow: 0px 0px 12px 0px rgba(0, 0, 0, 0.25);
-}
-
-.modal .loginModal .close {
-  float: right;
-  font-size: 25px;
-}
-
-.modal .loginModal .modalContents {
-  margin: 0 auto;
-  width: 100%;
-  position: relative;
-  padding: 0 20px 32px;
-  box-sizing: border-box;
-  display: flex;
-  justify-content: center;
-  flex-direction: column;
-}
-
-.modal .loginModal .modalContents .signinIcon {
-  width: 150px;
-  margin: 0 auto;
-}
-
-.modal .loginModal .modalContents .loginId {
-  margin-top: 30px;
-  border-radius: 2px;
-  width: 100%;
-  height: 40px;
-  border: 1px solid #e5e5e5;
-  padding: 9px 12px;
-  outline: none;
-  box-sizing: border-box;
-}
-
-.modal .loginModal .modalContents input::placeholder {
-  color: #999999;
-}
-
-.modal .loginModal .modalContents .loginPw {
-  margin-top: 15px;
-  border-radius: 2px;
-  width: 100%;
-  height: 40px;
-  border: 1px solid #e5e5e5;
-  padding: 9px 12px;
-  outline: none;
-  box-sizing: border-box;
-}
-
-.modal .loginModal .modalContents .loginMid {
-  display: flex;
-
-  justify-content: space-between;
-}
-
-.modal .loginModal .modalContents .loginMid .saveId {
-  font-size: 14px;
-  position: relative;
-  left: 30px;
-  color: #8d8d8d;
-  line-height: 3;
-  font-family: Noto Sans KR;
-  font-weight: 500;
-}
-
-.modal .loginModal .modalContents .loginMid #saveId {
-  margin-right: 10px;
-  cursor: pointer;
-}
-
-.modal .loginModal .modalContents .loginBtn {
-  width: 392px;
-  height: 56px;
-  font-family: Noto Sans KR;
-  font-size: 14px;
-  letter-spacing: -0.333px;
-  margin: 0 auto 30px;
-  background-color: green;
-  border: 0;
-}
-
-.loginRegisterBox {
-  display: flex;
-  flex-direction: column;
-  margin-top: 40px;
-}
-
-.register {
-  display: flex;
-  justify-content: center;
-}
-
-.register a {
-  color: #41a24b;
-  font-family: Noto Sans KR;
-  font-size: 14px;
-  letter-spacing: -0.333px;
-  text-decoration-line: underline;
-}
-
-.modal .loginModal .modalContents .socialBox {
-  margin-bottom: 30px;
-}
-
-.socialBox .socialHeader {
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  margin-top: 60px;
-}
-
-.socialHeader .line {
-  margin: 0 25px;
-  width: 98px;
-  height: 1px;
-  background: #e8e8e8;
-}
-
-.socialText {
-  margin: 30px 0;
-  color: #939393;
-  font-family: Noto Sans KR;
-  font-size: 14px;
-  letter-spacing: -0.333px;
-}
-
-.modal .loginModal .modalContents .socialBox .kakao {
-  background-color: #feec34;
-  border-color: #feec34;
-  margin: 0 auto 10px;
-  width: 393px;
-  height: 56px;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  box-sizing: border-box;
-  border-radius: 3px;
-  cursor: pointer;
-}
-
-.modal .loginModal .modalContents .socialBox .kakao .kakaoLogo {
-  width: 20px;
-  height: 20px;
-}
-
-.modal .loginModal .modalContents .socialBox .kakao .kakaoText {
-  width: 300px;
-  text-align: center;
-  display: inline-block;
-  box-sizing: border-box;
-  font-size: 14px;
-  font-weight: 500;
-}
-
-.modal .loginModal .modalContents .socialBox .google {
-  border-radius: 2px;
-  margin: 0 auto 10px;
-  width: 393px;
-  height: 56px;
-  border: 1px solid #e3e3e3;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  box-sizing: border-box;
-  color: #fff;
-  cursor: pointer;
-}
-
-.modal .loginModal .modalContents .socialBox .google .googleText {
-  width: 300px;
-  color: black;
-  text-align: center;
-  box-sizing: border-box;
-  font-size: 14px;
-  font-weight: 500;
-}
-
-.modal .loginModal .modalContents .socialBox .google .googleLogo {
-  width: 20px;
-  height: 20px;
-}
-
-.modal .loginModal .modalContents .loginEnd {
-  text-align: center;
-  font-size: 11px;
-}
-
-.modal .loginModal .modalContents .loginEnd .loginLine {
-  color: #bcbcbc;
-  font-size: 11px;
-  margin-bottom: 35px;
-}
-
-.modal .loginModal .modalContents .loginEnd .loginLine a {
-  color: black;
-  text-decoration: underline;
-  cursor: pointer;
-}
-
-.modal .loginModal .modalContents .loginEnd .noUser {
-  text-decoration: underline;
-}
-
-@media (max-width: 1024px) {
-  .modal .loginModal .modalContents .socialBox .socialHeader {
-    height: 60px;
-    margin: 0;
-  }
-  .modal .loginModal .modalContents .socialBox .socialHeader {
-    display: none;
-  }
-  .login-container h1 {
-    margin: 0;
-  }
-  .invalid-feedback {
-    height: 0px;
-    margin: 0;
-  }
-  .loginRegisterBox {
-    margin-top: 0px;
-  }
-  .register a {
-    margin-bottom: 20px;
-  }
-}
-@media (max-width: 767px) {
-  .modal .loginModal {
-    width: 100vw;
-  }
-  .login-container {
-    width: 80vw;
-  }
-  .login-container input {
-    width: 80vw;
-  }
-  .modal .loginModal .modalContents .loginBtn {
-    width: 80vw;
-  }
-
-  .modal .loginModal .modalContents .socialBox .kakao {
-    width: 80vw;
-    padding: 0 30px;
-  }
-  .modal .loginModal .modalContents .socialBox .google {
-    width: 80vw;
-    padding: 0 30px;
-  }
-  .find a {
-    font-size: 10px;
-  }
-}
-
-@media (max-width: 540px) {
-  .invalid-feedback {
-    height: 0px;
-    margin: 0;
-  }
-  .modal .loginModal .modalContents .socialBox .socialHeader {
-    height: 60px;
-    margin: 0;
-  }
-}
-
-@media (max-width: 375px) {
-  .invalid-feedback {
-    height: 0px;
-    margin: 0;
-  }
-  .loginRegisterBox {
-    margin-top: 0px;
-  }
-  .modal .loginModal .modalContents .socialBox .socialHeader {
-    height: 60px;
-    margin: 0;
-  }
-}
-
-@media (max-width: 320px) {
-  .modal .loginModal .modalContents .socialBox .socialHeader {
-    display: none;
-  }
-  .login-container h1 {
-    margin: 0;
-  }
-  .register a {
-    margin-bottom: 10px;
-  }
-}
-
-@media (max-width: 280px) {
-  .modal .loginModal .modalContents .socialBox .socialHeader {
-    display: none;
-  }
-  .register a {
-    margin-bottom: 30px;
-  }
-  .loginMid {
-    display: flex;
-    flex-direction: column;
-  }
-  .saveId {
-    height: 20px;
-  }
-}
