+++ src/app/components/SignUpAgreeModal/SignUpAgreeModal.css
@@ -179,6 +179,7 @@
 
 .btn-box {
   display: flex;
+  color: white;
 }
 
 @media (max-width: 1024px) {
